agents/requirements.txt
livekit-agents[openai,deepgram,cartesia,bey,hedra,silero,turn-detector,anam]>=1.2.2
Pillow



# -----------------------------
# Core
# -----------------------------
python-dotenv==1.2.1
Pillow==10.3.0
numpy==1.26.4
psutil==7.2.2

# -----------------------------
# LiveKit Core
# -----------------------------
livekit==1.0.23
livekit-agents==1.3.12
livekit-api==1.1.0
livekit-protocol==1.1.2
livekit-blingfire==1.1.0

# -----------------------------
# LiveKit Plugins
# -----------------------------
livekit-plugins-openai==1.3.12
livekit-plugins-silero==1.3.12
livekit-plugins-turn-detector==1.3.12
livekit-plugins-deepgram==1.3.12
livekit-plugins-cartesia==1.3.12
livekit-plugins-anam==1.3.12
livekit-plugins-bey==1.3.12
livekit-plugins-hedra==1.3.12

# -----------------------------
# Async / Networking
# -----------------------------
aiofiles==25.1.0
aiohttp==3.13.3
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
yarl==1.22.0
watchfiles==1.1.1

# -----------------------------
# Audio / Media
# -----------------------------
av==16.1.0
sounddevice==0.5.3
onnxruntime==1.23.1

# -----------------------------
# LLM / AI
# -----------------------------
openai==2.16.0
transformers==4.57.1
tokenizers==0.22.2
huggingface-hub==0.36.0
safetensors==0.7.0

# -----------------------------
# Telemetry (MATCHED PROTOBUF)
# -----------------------------
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-semantic-conventions==0.60b1

# -----------------------------
# Protobuf (CRITICAL PIN)
# -----------------------------
protobuf==6.33.4

# -----------------------------
# Utilities
# -----------------------------
typer==0.21.1
prometheus-client==0.24.1
types-protobuf==6.32.1.20251210
docstring-parser==0.17.0
eval-type-backport==0.3.1
websockets==15.0.1


# Mysql 
mysql-connector-python==9.6.0


email-validator==2.3.0